Reduce reducers
